Fix #54
To use with the RP2040 Mbed Arduino board package the User_Config.h file in the src folder must be eddited to comment out line 2 thus:
```
  //#define TJPGD_LOAD_FFS
```
This is required since the Mbed package does not support the LittleFS or SPIFFS FLASH filing system. Clearly the LittleFS and SPIFFS examples will not compile for this reason.